Try explicit long long

+++ tests/run/numpy_divmod.pyx
@@ -7,8 +7,8 @@ cpdef test_divmod():
     >>> test_divmod()
     -420000000
     """
-    cdef int64_t val = -420000000000
-    us, remainder = divmod(val, 1000LL)
+    cdef signed long long val = -420000000000
+    us, remainder = divmod(val, 1000)
     if remainder >= 500:
         us += 1
     return us
